[GRetriever] generate embeddings and test inference-only model

+++ implementation/sceneGraphsDataset.py
@@ -2,40 +2,23 @@ import torch
 import pandas as pd
 import json
 import re
+import os
 from tqdm import tqdm
 from torch.utils.data import Dataset
 from torch_geometric.data.data import Data
 from datasets import load_dataset, concatenate_datasets
 from textEmbedder import textEmbedder
-
-import sys
-
-def get_size(obj, seen=None):
-    """객체의 메모리 크기를 반환하는 함수"""
-    size = sys.getsizeof(obj)
-    if seen is None:
-        seen = set()
-    obj_id = id(obj)
-    if obj_id in seen:
-        return 0
-    # 객체의 ID를 seen 집합에 추가
-    seen.add(obj_id)
-    if isinstance(obj, dict):
-        size += sum([get_size(v, seen) for v in obj.values()])
-        size += sum([get_size(k, seen) for k in obj.keys()])
-    elif hasattr(obj, '__dict__'):
-        size += get_size(obj.__dict__, seen)
-    elif hasattr(obj, '__iter__') and not isinstance(obj, (str, bytes, bytearray)):
-        size += sum([get_size(i, seen) for i in obj])
-    return size
+from evaluate import eval_funcs
+from retriever import *
 
 IMAGE_PATH = "../data/sceneGraphs/sceneGraphs.json"
 QUESTION_PATH = "../data/sceneGraphs/sceneGraphsQuestions.csv"
-GRAPH_EMBEDDING_PATH = "../data/sceneGraphs/sceneGraphs.pt"
+GRAPH_EMBEDDING_PATH = "/mnt/sde/shcha/sceneGraphs.pt"
+RETRIEVED_GRAPH_EMBEDDING_PATH = "../data/sceneGraphs/retrievedSceneGraphs.pt"
 QUESTION_EMBEDDING_PATH = "../data/sceneGraphs/questionEmbs.pt"
 
 class SceneGraphsDataset(Dataset):
-    def __init__(self):
+    def __init__(self, useGR=False):
         super().__init__()
         with open(IMAGE_PATH, 'r', encoding='utf-8') as file:
             self.imageset = json.load(file)
@@ -44,33 +27,40 @@ class SceneGraphsDataset(Dataset):
         # nodes & edges are dictionary
         self.nodes, self.edges = self.extractNodesAndEdges()
         try:
-            self.graphEmbs = torch.load(GRAPH_EMBEDDING_PATH, weights_only=False)
+            print("Loading sceneGraphs' graph embeddings ...")
+            graphPath = RETRIEVED_GRAPH_EMBEDDING_PATH if useGR else GRAPH_EMBEDDING_PATH
+            self.graphEmbs = torch.load(graphPath, weights_only=False)
         except:
             print("Fail to load sceneGraphs' graph embeddings.")
             self.graphEmbs = []
         try:
+            print("Loading sceneGraphs' question embeddings ...")
             self.questionEmbs = torch.load(QUESTION_EMBEDDING_PATH, weights_only=False)
         except:
             print("Fail to load sceneGraphs' question embeddings.")
             self.questionEmbs = []
 
     def __len__(self):
-        return len(self.questionset['label'])
+        return len(self.questionset['answer'])
     
     def __getitem__(self, index):
-        assert len(self.questionset['label']) > index and len(self.graphEmbs) > index and len(self.nodes) > index and len(self.edges) > index
+        assert len(self.questionset['answer']) > index
         imgIdx = self.questionset['image_id'][index]
-
+        nodesDf = pd.DataFrame(enumerate(self.nodes[imgIdx]), columns=['node_id', 'node_attr'])
+        edgesDf = pd.DataFrame(self.edges[imgIdx])
+        graphEmbs = self.graphEmbs[list(self.questionset['image_id'].unique()).index(imgIdx)]
+        #subg, desc = retrieval_via_pcst(graphEmbs, self.questionEmbs[index], nodesDf, edgesDf, topk=3, topk_e=3, cost_e=0.5)
         return {
             'index': index,
             'imageIdx': imgIdx,
-            'question': f'Question: {self.questionset['question'][index]}\nAnswer: ',
+            'question': f"Question: {self.questionset['question'][index]}\n\nAnswer: ",
             'label': self.questionset['answer'][index],
             'fullLabel': self.questionset['full_answer'][index],
-            'graphEmbs': self.graphEmbs[index],
+            'graphEmbs': graphEmbs,
             'qEmbs': self.questionEmbs[index],
-            'nodes': self.nodes[imgIdx],
-            'edges': self.edges[imgIdx]
+            #'desc': desc
+            #'desc': nodesDf.to_csv(index=False)+'\n'+edgesDf.to_csv(index=False)
+            'desc': nodesDf.to_csv(index=False).replace('\n','\n\n')+'\n'+edgesDf.to_csv(index=False).replace('\n','\n\n')
         }
 
     def extractNodesAndEdges(self):
@@ -90,30 +80,31 @@ class SceneGraphsDataset(Dataset):
                 for relation in object['relations']:
                     edges.append({'src': oid2nid[objectId], 'edge': relation['name'], 'dst': oid2nid[relation['object']]})
 
-            nodesDict[imageId] = nodes
-            edgesDict[imageId] = edges
+            nodesDict[int(imageId)] = nodes
+            edgesDict[int(imageId)] = edges
 
         return nodesDict, edgesDict
 
     def preprocessing(self):
         textEmbedder.loadModel['sbert']()
         text2embs = textEmbedder.runModel['sbert']
-        print("SceneGraphs preprocessing(Question) ... ")
-        processedQuestions = text2embs(self.questionset['question'])
-        torch.save(processedQuestions, QUESTION_EMBEDDING_PATH)
+        if not os.path.exists(QUESTION_EMBEDDING_PATH):
+            print("SceneGraphs preprocessing(Question) ... ")
+            processedQuestions = text2embs(self.questionset['question'].tolist())
+            torch.save(processedQuestions, QUESTION_EMBEDDING_PATH)
 
-        print("SceneGraphs preprocessing(Graph) ... ")
-        processedGraphs = []
-        for imgId in tqdm(self.questionset['image_id'].unique()):
-            nodeEmbs = text2embs(self.nodes[imgId])
-            edgeEmbs = text2embs([e['edge'] for e in self.edges[imgId]])
-            edgeIdx = torch.LongTensor([[e['src'] for e in self.edges[imgId]], 
-                                        [e['dst'] for e in self.edges[imgId]]])
-            processedGraph = Data(x=nodeEmbs, edge_index=edgeIdx, edge_attr=edgeEmbs, num_nodes=len(self.nodes[imgId]))
-            processedGraphs.append(processedGraph)
-            print(get_size(processedGraphs))
+        if not os.path.exists(GRAPH_EMBEDDING_PATH):
+            print("SceneGraphs preprocessing(Graph) ... ")
+            processedGraphs = []
+            for imgId in tqdm(self.questionset['image_id'].unique()):
+                nodeEmbs = text2embs(self.nodes[imgId])
+                edgeEmbs = text2embs([e['edge'] for e in self.edges[imgId]])
+                edgeIdx = torch.LongTensor([[e['src'] for e in self.edges[imgId]], 
+                                            [e['dst'] for e in self.edges[imgId]]])
+                processedGraph = Data(x=nodeEmbs, edge_index=edgeIdx, edge_attr=edgeEmbs, num_nodes=len(self.nodes[imgId]))
+                processedGraphs.append(processedGraph)
 
-        torch.save(processedGraphs, GRAPH_EMBEDDING_PATH)
+            torch.save(processedGraphs, GRAPH_EMBEDDING_PATH)
 
     def splitDataset(self):
         trainTail = self.__len__() * 6 // 10
@@ -124,5 +115,5 @@ class SceneGraphsDataset(Dataset):
         testIdxs = list(range(validationTail, testTail))
         return trainIdxs, validationIdxs, testIdxs
 
-sg = SceneGraphsDataset()
-sg.preprocessing()
+    def eval(self, path):
+        return eval_funcs['scene_graphs'](path)
